requirements.txt
asgiref==3.7.2
boto3==1.33.11
botocore==1.33.11
django-cors-headers==4.3.1
django-storages==1.14.2
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
packaging==23.2
jmespath==1.0.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
urllib3>=1.25.4,<1.27
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
dotenv==0.0.5